refactor: move duplicated connection methods to FakeBaseConnectionMixin
The sync (FakeBaseConnection) and async (FakeBaseAsyncConnection)
connection classes copy-pasted four identical methods: _decode,
repr_pieces, __str__, and _add_to_local_cache. Move them to their
shared base FakeBaseConnectionMixin in _server.py.

_get_from_local_cache stays in each subclass because the async client
requires it to be a coroutine; read_response stays because the
blocking/queue semantics genuinely differ.

No behavior change. mypy error count drops from 58 to 55.
